+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Added
+* HTTPS capability for the dev server ([#70](https://github.com/koopjs/koop-cli/issues/70))
+
 ## 1.0.0 - 2020-12-15
 ### Added
 * Plugin list in the `koop.json`